+++ src/applications/item-pile-config/item-pile-config.svelte
@@ -114,6 +114,10 @@
       }
     });
 
+    if (currentData.enabled && currentData.type === CONSTANTS.PILE_TYPES.VAULT && currentData.logVaultActions) {
+      await PileUtilities.getVaultJournalPageForItemPile(pileActor, { create: true });
+    }
+
     application.close();
 
   }